Fix:安全告警清理

+++ frameworks/core/components_ng/pattern/image/image_pattern.cpp
@@ -1716,7 +1716,7 @@ std::list<ImagePattern::CacheImageStruct>::iterator ImagePattern::FindCacheImage
 void ImagePattern::GenerateCachedImages()
 {
     CHECK_NULL_VOID(images_.size());
-    auto averageShowTime = animator_->GetDuration() / images_.size();
+    auto averageShowTime = static_cast<uint32_t>(animator_->GetDuration()) / images_.size();
     size_t cacheImageNum = averageShowTime >= CRITICAL_TIME ? 1 : 2;
     cacheImageNum = std::min(images_.size() - 1, cacheImageNum);
     if (cacheImages_.size() > cacheImageNum) {
